move decimalInput.jsx out of utils folder
